v1.0 - Super Mega Awesome Update
This brings around any outstanding functionality, and users can actually
get their recs and get links to a google search for torrentz of that
album. Going to have to do some major cleanup in terms of processing, as
well I would like to bring a lot more functionality to this thing but
for now this is a v1.0 release.