v1.0.2: Some UI Changes, A Devlog, & Some Code Documentation
Yep. Overkill on the media queries for the 'article.content' width, but
it makes for a smooth UX at any resolution. The really tiny mobile
screens kinda suck but like who's trying to find torrents on their
mobile device? Regardless I did that.

Would like to drop a v1.1 by the end of the day, with at least 2 or 3 of
the items on the Feature Request list dealt with.