v1.0.4 - Fixed all auth/no artists showing bugs; Commas in playcounts; Socials;
v1.0.4

Getting ready for a v1.1 release, this tackled the stupid bugs where no
artist was showing up, as well as the username not showing up.

	+ Artists are fetched after authing (like how they sometimes
were just getting added to localStorage but $scope.artists was never
updated.
	+ Username always shows up if it exists
	+ Commas every 3 digits in playcounts
	+ Social btns at the bottom of the page (neeet)

There will be one last pre-release before v1.1, which will be adding a
build task to the gulpfile. The 1.1 will then be made live.